20th commit

+++ admin-client/src/components/Courses.jsx
@@ -1,8 +1,10 @@
 import { Button, Card, Typography } from "@mui/material";
 import { useEffect, useState } from "react";
-import { useNavigate } from "react-router-dom";
 import { BASE_URL } from "../config.js";
 import axios from "axios";
+import { userEmailState } from "../store/selectors/userEmail.js";
+import { useRecoilValue } from "recoil";
+import { Loading } from "./Loading.jsx";
 
 function Courses() {
   const [courses, setCourses] = useState([]);
@@ -20,6 +22,14 @@ function Courses() {
     init();
   }, []);
 
+  if (courses.length == 0) {
+    return (
+      <div>
+        <Loading />
+      </div>
+    );
+  }
+
   return (
     <div
       style={{ display: "flex", flexWrap: "wrap", justifyContent: "center" }}
@@ -34,7 +44,7 @@ function Courses() {
 }
 
 export function Course({ course, setCourses, courses }) {
-  const navigate = useNavigate();
+  const userEmail = useRecoilValue(userEmailState);
   return (
     <Card
       style={{
@@ -73,21 +83,25 @@ export function Course({ course, setCourses, courses }) {
           size="large"
           style={{ marginTop: "10px" }}
           onClick={async () => {
-            try {
-              const response = await axios.delete(
-                `${BASE_URL}/admin/delete/${course._id}`,
-                {
-                  headers: {
-                    Authorization: `Bearer ${localStorage.getItem("token")}`,
-                  },
-                }
-              );
-              alert(response.data.message);
-              setCourses(
-                courses.filter((course1) => course1._id !== course._id)
-              );
-            } catch (err) {
-              alert(err.response.data.message);
+            if (userEmail == "faizaan") {
+              alert("you can't delete with default credentials!");
+            } else {
+              try {
+                const response = await axios.delete(
+                  `${BASE_URL}/admin/delete/${course._id}`,
+                  {
+                    headers: {
+                      Authorization: `Bearer ${localStorage.getItem("token")}`,
+                    },
+                  }
+                );
+                alert(response.data.message);
+                setCourses(
+                  courses.filter((course1) => course1._id !== course._id)
+                );
+              } catch (err) {
+                alert(err.response.data.message);
+              }
             }
           }}
         >
